requirement.txt
ffmpy3==0.2.3
numpy==1.20.1
playsound==1.2.2
PyAudio==0.2.11
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
six==1.15.0
websocket-client==0.58.0
